third_party/WebKit/LayoutTests/svg/animations/svgnumberoptionalnumber-animation-3-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

Test 'to' animation of SVGNumberOptionalNumber without optional number.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS feGaussianBlur.stdDeviationX.animVal is 40
PASS feGaussianBlur.stdDeviationY.animVal is 10
PASS feGaussianBlur.stdDeviationX.baseVal is 40
PASS feGaussianBlur.stdDeviationY.baseVal is 10
PASS feGaussianBlur.stdDeviationX.animVal is 22.5
PASS feGaussianBlur.stdDeviationY.animVal is 7.5
PASS feGaussianBlur.stdDeviationX.baseVal is 40
PASS feGaussianBlur.stdDeviationY.baseVal is 10
PASS feGaussianBlur.stdDeviationX.animVal is 5
PASS feGaussianBlur.stdDeviationY.animVal is 5
PASS feGaussianBlur.stdDeviationX.baseVal is 40
PASS feGaussianBlur.stdDeviationY.baseVal is 10
PASS feGaussianBlur.stdDeviationX.animVal is 40
PASS feGaussianBlur.stdDeviationY.animVal is 10
PASS feGaussianBlur.stdDeviationX.baseVal is 40
PASS feGaussianBlur.stdDeviationY.baseVal is 10
PASS successfullyParsed is true

TEST COMPLETE
